backend/pyproject.toml
[project]
name = "hanabibackend"
version = "0.1.0"
description = "Backend in fastAPI for hanabi"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.balck.dependencies]
isort = "^6.0.1"
flake8 = "^7.2.0"
ruff = "^0.11.11"
pre-commit = "^4.2.0"

